requirements.txt
celery
django
djangorestframework
djangorestframework-camel-case
django-anymail
django-cors-headers
django-crispy-forms
django_extensions
django-filter
dj-database-url
factory_boy
ipython
psycopg2
requests
uwsgi
pillow
python-dateutil
